Add Google Custom Search API keys and update package references, Refactored

+++ Library/Forms/FormStaff.cs
@@ -10,6 +10,7 @@ using System.Threading.Tasks;
 using System.Windows.Forms;
 using Microsoft.EntityFrameworkCore;
 using Library.Components;
+using Library.Components.FromStaff;
 using static Library.Forms.FromCopies;
 using static System.Windows.Forms.VisualStyles.VisualStyleElement;
 
@@ -20,13 +21,7 @@ namespace Library.Forms
         public FormStaff()
         {
             InitializeComponent();
-
-            dataGridView1.ColumnHeadersDefaultCellStyle.BackColor = Color.DarkBlue;
-            dataGridView1.ColumnHeadersDefaultCellStyle.ForeColor = Color.White;
-            dataGridView1.ColumnHeadersDefaultCellStyle.Font = new Font("Segoe UI", 10, FontStyle.Bold);
-            dataGridView1.DefaultCellStyle.SelectionBackColor = Color.Silver;
-            dataGridView1.DefaultCellStyle.SelectionForeColor = Color.Black;
-            dataGridView1.AlternatingRowsDefaultCellStyle.BackColor = Color.LightGray;
+            DataGridStyle.DefaultStyle(dataGridView1);
 
             var binder = new DataGridBinder();
             binder.BindDataToGrid(
@@ -45,65 +40,26 @@ namespace Library.Forms
             );
         }
 
-        public class StaffViewModel
-        {
-            public int StaffId { get; set; }
-            public int BranchId { get; set; }
-            public string BranchName { get; set; }
-            public string StaffName { get; set; }
-            public string Possition { get; set; }
-            public double Salary { get; set; }
-        }
-
+        // handles staff selection
         private void dataGridView1_SelectionChanged(object sender, EventArgs e)
         {
-            if (dataGridView1.SelectedRows.Count > 0)
-            {
-                BindingList<StaffViewModel>? data = dataGridView1.DataSource as BindingList<StaffViewModel>;
-                if (data != null)
-                {
-                    var selectedRowIndex = dataGridView1.SelectedRows[0].Index;
-                    if (selectedRowIndex >= 0 && selectedRowIndex < data.Count)
-                    {
-                        var selectedStaff = data[selectedRowIndex];
-                        textBox1.Text = selectedStaff.Salary.ToString();
-                    }
-                }
-            }
+            StaffSelection.HandleStaffSelection(dataGridView1, textBox1);
         }
 
+        // deletes selected row
         private void btnDelete_Click(object sender, EventArgs e)
         {
+            DialogResult dialogResult = MessageBox.Show("Are you sure you want to delete staff?", "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+            if (dialogResult != DialogResult.Yes) return;
+
             var deleteDataGridRow = new DeleteDataGridRow();
             deleteDataGridRow.DeleteSelectedRow<Staff, StaffViewModel>(dataGridView1, viewModel => viewModel.StaffId);
-            
         }
 
-
+        // changes salary of selected staff
         private void button1_Click(object sender, EventArgs e)
         {
-            if (dataGridView1.SelectedRows.Count > 0)
-            {
-                BindingList<StaffViewModel>? data = dataGridView1.DataSource as BindingList<StaffViewModel>;
-                if (data != null)
-                {
-                    var selectedRowIndex = dataGridView1.SelectedRows[0].Index;
-                    if (selectedRowIndex >= 0 && selectedRowIndex < data.Count)
-                    {
-                        var selectedStaff = data[selectedRowIndex];
-                        double newSalary;
-                        if (double.TryParse(textBox1.Text, out newSalary))
-                        {
-                            selectedStaff.Salary = newSalary;
-                            dataGridView1.Refresh();
-                        }
-                        else
-                        {
-                            //error
-                        }
-                    }
-                }
-            }
+            StaffEdit.HandleStaffSalaryChange(dataGridView1, textBox1);
         }
     }
 }
